Time: 3 ms (66.58%), Space: 9 MB (64.81%) - LeetHub

+++ 1431-kids-with-the-greatest-number-of-candies/1431-kids-with-the-greatest-number-of-candies.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     vector<bool> kidsWithCandies(vector<int>& candies, int extraCandies) {
-        int n=candies.size();
+        short n=candies.size();
         
         vector<bool>ans(n);
         
-        int mx=0;
+        short mx=0;
         
-        for(auto it:candies)mx=max(mx,it);
+        for(auto it:candies)mx=max(mx,(short)it);
         
         for(int i=0;i<n;i++)ans[i]=(candies[i]+extraCandies)>=mx;
         return ans;
